+++ frontend/src/core/components/viewer/BookmarkSidebar.css
@@ -1,63 +1,4 @@
-/* Bookmark Sidebar - Modern styling to match AllTools section */
-
-.bookmark-sidebar {
-  display: flex;
-  flex-direction: column;
-  background: linear-gradient(
-    135deg,
-    color-mix(in srgb, var(--bg-toolbar) 96%, transparent),
-    color-mix(in srgb, var(--bg-background) 90%, transparent)
-  );
-  border-left: 1px solid color-mix(in srgb, var(--border-subtle) 75%, transparent);
-  box-shadow: -2px 0 16px color-mix(in srgb, var(--shadow-color, rgba(15, 23, 42, 0.35)) 20%, transparent);
-  backdrop-filter: blur(12px);
-}
-
-/* Header Section */
-.bookmark-sidebar__header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0.75rem 0.875rem;
-  background: var(--bg-toolbar);
-  border-bottom: 1px solid var(--border-subtle);
-}
-
-.bookmark-sidebar__header-title {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.bookmark-sidebar__header-icon {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: var(--mantine-color-blue-6);
-  font-size: 1.1rem;
-}
-
-/* Search Section */
-.bookmark-sidebar__search {
-  background: var(--tool-panel-search-bg, var(--bg-toolbar));
-  border-bottom: 1px solid var(--tool-panel-search-border-bottom, var(--border-subtle));
-  padding-top: 0.75rem !important;
-}
-
-/* Content Section */
-.bookmark-sidebar__content {
-  padding-top: 0.5rem !important;
-}
-
-.bookmark-sidebar__empty-state,
-.bookmark-sidebar__error {
-  padding: 2rem 1rem;
-  text-align: center;
-}
-
-.bookmark-sidebar__loading {
-  padding: 2rem 1rem !important;
-}
+/* Bookmark Sidebar - Bookmark-specific styles (base styles in SidebarBase.css) */
 
 /* Bookmark List */
 .bookmark-list {
@@ -193,4 +134,3 @@
     animation: none;
   }
 }
-
