+++ frontend/src/core/components/viewer/LayerSidebar.css
@@ -1,65 +1,11 @@
-/* Layer Sidebar - Styling consistent with BookmarkSidebar */
+/* Layer Sidebar - Layer-specific styles (base styles in SidebarBase.css) */
 
-.layer-sidebar {
-  display: flex;
-  flex-direction: column;
-  background: linear-gradient(
-    135deg,
-    color-mix(in srgb, var(--bg-toolbar) 96%, transparent),
-    color-mix(in srgb, var(--bg-background) 90%, transparent)
-  );
-  border-left: 1px solid color-mix(in srgb, var(--border-subtle) 75%, transparent);
-  box-shadow: -2px 0 16px color-mix(in srgb, var(--shadow-color, rgba(15, 23, 42, 0.35)) 20%, transparent);
-  backdrop-filter: blur(12px);
-}
-
-/* Header Section */
-.layer-sidebar__header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0.75rem 0.875rem;
-  background: var(--bg-toolbar);
-  border-bottom: 1px solid var(--border-subtle);
-  gap: 0.5rem;
-}
-
-.layer-sidebar__header-title {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  min-width: 0;
-}
-
-.layer-sidebar__header-icon {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: var(--mantine-color-blue-6);
-  font-size: 1.1rem;
-  flex-shrink: 0;
-}
-
+/* Header actions (show/hide all buttons) */
 .layer-sidebar__header-actions {
   display: flex;
   align-items: center;
   gap: 0.25rem;
   flex-shrink: 0;
-}
-
-/* Content Section */
-.layer-sidebar__content {
-  padding-top: 0.5rem !important;
-}
-
-.layer-sidebar__empty-state,
-.layer-sidebar__error {
-  padding: 2rem 1rem;
-  text-align: center;
-}
-
-.layer-sidebar__loading {
-  padding: 2rem 1rem !important;
 }
 
 /* Layer List */
